Startup logo in runtime.
